requirements.txt
asgiref==3.5.2
bidict==0.22.0
cffi==1.15.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
enum-compat==0.0.3
eventlet==0.33.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
pycparser==2.21
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
zope.event==4.5.0
zope.interface==5.5.2
